+++ Samples/iOS-ObjectiveC/iOS-ObjectiveC/ViewController.m
@@ -103,6 +103,7 @@ ViewController ()
 - (IBAction)captureTransaction:(id)sender
 {
     SentryTransaction *fakeTransaction = [SentrySDK startTransactionWithName:@"Some Transaction"];
+
     dispatch_after(
         dispatch_time(DISPATCH_TIME_NOW, (int64_t)(arc4random_uniform(100) + 400 * NSEC_PER_MSEC)),
         dispatch_get_main_queue(), ^{ [fakeTransaction finish]; });