+++ Tests/SentryTests/SentryTransactionTests.swift
@@ -2,28 +2,67 @@ import XCTest
 
 class SentryTransactionTest: XCTestCase {
     let someTransactionName = "Some Transaction"
+    let someOperation = "Some Operation"
     
     func testInitWithName() {
-        let transaction = SentryTransaction(name: someTransactionName)
+        let transaction = Transaction(name: someTransactionName)
+      
         XCTAssertNotNil(transaction.startTimestamp)
         XCTAssertNil(transaction.timestamp)
         XCTAssertEqual(transaction.transaction, someTransactionName)
     }
     
     func testInitWithTransactionContext() {
-        let context = TransactionContext(name: someTransactionName)
-        let transaction = SentryTransaction(transactionContext: context, andHub: nil)
+        let someOperation = "Some Operation"
+        let someSpanDescription = "Some Span Description"
+        
+        let context = TransactionContext(name: someTransactionName, trace: SentryId(), spanId: SpanId(), parentSpanId: SpanId(), andParentSampled: true)
+        context.operation = someOperation
+        context.status = .ok
+        context.sampled = true
+        context.spanDescription = someSpanDescription
+        
+        let transaction = Transaction(transactionContext: context, andHub: nil)
         XCTAssertNotNil(transaction.startTimestamp)
         XCTAssertNil(transaction.timestamp)
         XCTAssertEqual(transaction.transaction, someTransactionName)
+        XCTAssertEqual(transaction.traceId, context.traceId)
+        XCTAssertEqual(transaction.spanId, context.spanId)
+        XCTAssertEqual(transaction.operation, someOperation)
+        XCTAssertEqual(transaction.status, SentrySpanStatus.ok)
+        XCTAssertTrue(transaction.isSampled)
+        XCTAssertEqual(transaction.spanDescription, someSpanDescription)
+    }
+    
+    func testIndirectManipulationOfContext() {
+        let someOperation = "Some Operation"
+        let spanDescription = "Span Description"
+        
+        let context = TransactionContext(name: someTransactionName)
+        
+        let transaction = Transaction(transactionContext: context, andHub: nil)
+        transaction.spanDescription = spanDescription
+        transaction.operation = someOperation
+        transaction.status = .ok
+        
+        XCTAssertEqual(context.spanDescription, spanDescription)
+        XCTAssertEqual(context.operation, someOperation)
+        XCTAssertEqual(context.status, SentrySpanStatus.ok)
     }
     
     func testInitWithNameAndContext() {
         let context = SpanContext()
-        let transaction = SentryTransaction(name: someTransactionName, spanContext: context, andHub: nil)
+        context.operation = someOperation
+        context.status = .ok
+
+        let transaction = Transaction(name: someTransactionName, spanContext: context, andHub: nil)
         XCTAssertNotNil(transaction.startTimestamp)
         XCTAssertNil(transaction.timestamp)
         XCTAssertEqual(transaction.transaction, someTransactionName)
+        XCTAssertEqual(transaction.traceId, context.traceId)
+        XCTAssertEqual(transaction.spanId, context.spanId)
+        XCTAssertEqual(transaction.operation, someOperation)
+        XCTAssertEqual(transaction.status, SentrySpanStatus.ok)
     }
     
     func testFinishCapturesTransaction() {
@@ -32,7 +71,7 @@ class SentryTransactionTest: XCTestCase {
         let client = TestClient(options: Options(), andTransport: transport, andFileManager: fileManager)
         let hub = SentryHub(client: client, andScope: nil, andCrashAdapter: TestSentryCrashWrapper())
 
-        let transaction = SentryTransaction(name: someTransactionName, spanContext: SpanContext(), andHub: hub)
+        let transaction = Transaction(name: someTransactionName, spanContext: SpanContext(), andHub: hub)
         transaction.finish()
 
         XCTAssertNotNil(transaction.startTimestamp)
@@ -40,17 +79,35 @@ class SentryTransactionTest: XCTestCase {
         XCTAssertTrue(transaction.timestamp! >= transaction.startTimestamp!)
         XCTAssertTrue(client.captureEventWithScopeArguments.last!.event === transaction)
     }
-    
-    func testSerialization() {
-        let transaction = SentryTransaction(name: someTransactionName)
-        transaction.finish()
+
+    func testSerializationWithoutContext() {
+        let transaction = Transaction(name: someTransactionName)
         
         let serialization = transaction.serialize()
+        XCTAssertNotNil(serialization)
         XCTAssertEqual(serialization["type"] as? String, "transaction")
         XCTAssertNotNil(serialization["event_id"])
         XCTAssertNotNil(serialization["start_timestamp"])
         XCTAssertNotNil(serialization["timestamp"])
         XCTAssertEqual(serialization["transaction"] as? String, someTransactionName)
         XCTAssertNotNil(serialization["contexts"])
+        XCTAssertNotNil((serialization["contexts"] as! Dictionary)["trace"])
+        XCTAssertNotNil(serialization["spans"])
+    }
+    
+    func testSerializationWithContext() {
+        let transaction = Transaction(name: someTransactionName)
+        transaction.context = [String: [String: Any]]()
+        
+        let serialization = transaction.serialize()
+        XCTAssertNotNil(serialization)
+        XCTAssertEqual(serialization["type"] as? String, "transaction")
+        XCTAssertNotNil(serialization["event_id"])
+        XCTAssertNotNil(serialization["start_timestamp"])
+        XCTAssertNotNil(serialization["timestamp"])
+        XCTAssertEqual(serialization["transaction"] as? String, someTransactionName)
+        XCTAssertNotNil(serialization["contexts"])
+        XCTAssertNotNil((serialization["contexts"] as! Dictionary)["trace"])
+        XCTAssertNotNil(serialization["spans"])
     }
 }