+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## unreleased
 
+- feat: Expand SentrySpanContext through SentryTransaction #919 
+- feat: Performance Monitoring API
 - feat: Add sendDefaultPii to SentryOptions #923
 
-- feat: Performance Monitoring API
 ## 6.1.4
 
 - fix: Sessions for Hybrid SDKs #913
-- feat: Performance Monitoring API
 
 ## 6.1.3